+++ docs/changelog.rst
@@ -12,8 +12,8 @@ The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`
   :local:
   :depth: 1
 
-Unreleased
-==========
+Unreleased (probably v0.5)
+==========================
 
 Added
 -----
@@ -23,10 +23,12 @@ Added
 - :class:`~gridsim.logger.Logger` now supports logging dictionary parameters (including those saved from config files).
 - When logging parameters with :meth:`~gridsim.logger.Logger.log_param`, you can now specify a subgroup of ``params`` in which to save the data.
 - :class:`~gridsim.message.Message` now has a method :meth:`~gridsim.message.Message.set_all` to replace the whole contents of the Message without creating a new message. (And as opposed to setting the contents key-by-key with :meth:`~gridsim.message.Message.set_all`.)
+- Documentation for profiling code (under Development)
 
 Changed
 -------
 
+- :meth:`~gridsim.robot.Robot.comm_criteria` and :meth:`~gridsim.robot.Robot.receive_msg` now have an argument of `dist_sqr` instead of `dist`. This is because the square root call in computing the distance is computationally expensive and is called for communication (an N^2) operation.
 - For :meth:`~gridsim.world.World.tag`, you can now pass ``None`` instead of a color to remove an existing tag.
 - [Under the hood] Convert World tagging to use 3D numpy array instead of list of Tuples. This also improves drawing speed ( it doesn't slow down so much when more tags are added), and you don't get gaps between tagged cells for certain window sizes.
 - Improved code documentation.